refactor(settings): extract keybindings subsystem into extension
Lifts the keybindings/shortcuts code (~124 lines) into
BellithSettings+Keybindings.swift: keybindings property, shortcut/
binding lookup helpers, conflict detection, reset(actionId:) /
reset(category:), applyPreset(_:), activeDefaultKeybindings,
mergeBinding, defaultKeybindings static, legacyPaneActionIDs.

mergeBinding and activeDefaultKeybindings bump from private to internal
to support the cross-file extension.

BellithSettings.swift: 862 → 740 lines.
